Extract getAxisConfig into a helper file
